Rename folder schoolPages to school
