+++ Gladiator.Presentation.Ui/src/pages/school/SchoolDetailsPage.js
