+++ Gladiator.Presentation.Ui/src/pages/school/SchoolListPage.js
